+++ src/shared/ui/button/index.ts
@@ -1,0 +1,5 @@
+export {ContainedButton} from '@/shared/ui/button/ContainedButton.tsx'
+
+export {IconButton} from '@/shared/ui/button/IconButton.tsx';
+
+export {OutlinedButton} from '@/shared/ui/button/OutlinedButton.tsx'
